WIP Start refactoring separation of concerns for repo metadata
